+++ python/tests/test_graph_conversions.py
@@ -953,36 +953,36 @@ def save_df_to_json(df, filename):
 def build_to_df():
     g = build_graph()
 
-    edge_df = g.to_edge_df()
+    edge_df = g..edges.to_df()
     save_df_to_json(edge_df, "expected/dataframe_output/edge_df_all.json")
-    edge_df = g.to_edge_df(include_property_histories=False)
+    edge_df = g..edges.to_df(include_property_histories=False)
     save_df_to_json(edge_df, "expected/dataframe_output/edge_df_no_props.json")
-    edge_df = g.to_edge_df(explode=False)
+    edge_df = g..edges.to_df(explode=False)
     save_df_to_json(edge_df, "expected/dataframe_output/edge_df_no_explode.json")
-    edge_df = g.to_edge_df(convert_datetime=False)
+    edge_df = g..edges.to_df(convert_datetime=False)
     save_df_to_json(edge_df, "expected/dataframe_output/edge_df_no_datetime.json")
 
-    edge_df = g.to_edge_df(explode=True)
+    edge_df = g..edges.to_df(explode=True)
     save_df_to_json(edge_df, "expected/dataframe_output/edge_df_explode.json")
-    edge_df = g.to_edge_df(explode=True, convert_datetime=True)
+    edge_df = g..edges.to_df(explode=True, convert_datetime=True)
     save_df_to_json(edge_df, "expected/dataframe_output/edge_df_explode_datetime.json")
-    edge_df = g.to_edge_df(explode=True, include_property_histories=True)
+    edge_df = g..edges.to_df(explode=True, include_property_histories=True)
     save_df_to_json(edge_df, "expected/dataframe_output/edge_df_explode_no_hist.json")
 
-    node_df = g.to_node_df()
+    node_df = g.nodes.to_df()
     save_df_to_json(node_df, "expected/dataframe_output/node_df_all.json")
-    node_df = g.to_node_df(include_property_histories=False)
+    node_df = g.nodes.to_df(include_property_histories=False)
     save_df_to_json(node_df, "expected/dataframe_output/node_df_no_hist.json")
-    node_df = g.to_node_df(explode=False)
+    node_df = g.nodes.to_df(explode=False)
     save_df_to_json(node_df, "expected/dataframe_output/node_df_no_explode.json")
-    node_df = g.to_node_df(convert_datetime=False)
+    node_df = g.nodes.to_df(convert_datetime=False)
     save_df_to_json(node_df, "expected/dataframe_output/node_df_no_datetime.json")
 
-    node_df = g.to_node_df(explode=True)
+    node_df = g.nodes.to_df(explode=True)
     save_df_to_json(node_df, "expected/dataframe_output/node_df_explode.json")
-    node_df = g.to_node_df(explode=True, include_property_histories=False)
+    node_df = g.nodes.to_df(explode=True, include_property_histories=False)
     save_df_to_json(node_df, "expected/dataframe_output/node_df_explode_no_hist.json")
-    node_df = g.to_node_df(explode=True, convert_datetime=True)
+    node_df = g.nodes.to_df(explode=True, convert_datetime=True)
     save_df_to_json(node_df, "expected/dataframe_output/node_df_explode_datetime.json")
 
 
@@ -1021,70 +1021,70 @@ def test_to_df():
     g = build_graph()
 
     compare_df(
-        g.to_edge_df(),
+        g..edges.to_df(),
         pd.read_json(base_dir / "expected/dataframe_output/edge_df_all.json"),
     )
 
     compare_df(
-        g.to_edge_df(include_property_histories=False),
+        g..edges.to_df(include_property_histories=False),
         pd.read_json(base_dir / "expected/dataframe_output/edge_df_no_props.json"),
     )
 
     compare_df(
-        g.to_edge_df(explode=False),
+        g..edges.to_df(explode=False),
         pd.read_json(base_dir / "expected/dataframe_output/edge_df_no_explode.json"),
     )
 
     compare_df(
-        g.to_edge_df(convert_datetime=False),
+        g..edges.to_df(convert_datetime=False),
         pd.read_json(base_dir / "expected/dataframe_output/edge_df_no_datetime.json"),
     )
 
     compare_df(
-        g.to_edge_df(explode=True),
+        g..edges.to_df(explode=True),
         pd.read_json(base_dir / "expected/dataframe_output/edge_df_explode.json"),
     )
 
     # compare_df(
-    #     g.to_edge_df(explode=True, convert_datetime=True),
+    #     g..edges.to_df(explode=True, convert_datetime=True),
     #     pd.read_json(base_dir / "expected/dataframe_output/edge_df_explode_datetime.json")
     # )
 
     compare_df(
-        g.to_edge_df(explode=True, include_property_histories=True),
+        g..edges.to_df(explode=True, include_property_histories=True),
         pd.read_json(
             base_dir / "expected/dataframe_output/edge_df_explode_no_hist.json"
         ),
     )
 
     compare_df(
-        g.to_node_df(include_property_histories=False),
+        g.nodes.to_df(include_property_histories=False),
         pd.read_json(base_dir / "expected/dataframe_output/node_df_no_hist.json"),
     )
 
     compare_df(
-        g.to_node_df(explode=False, include_property_histories=True),
+        g.nodes.to_df(explode=False, include_property_histories=True),
         pd.read_json(base_dir / "expected/dataframe_output/node_df_no_explode.json"),
     )
 
     compare_df(
-        g.to_node_df(convert_datetime=False, include_property_histories=True),
+        g.nodes.to_df(convert_datetime=False, include_property_histories=True),
         pd.read_json(base_dir / "expected/dataframe_output/node_df_no_datetime.json"),
     )
 
     compare_df(
-        g.to_node_df(explode=True),
+        g.nodes.to_df(explode=True),
         pd.read_json(base_dir / "expected/dataframe_output/node_df_explode.json"),
     )
 
     compare_df(
-        g.to_node_df(explode=True, include_property_histories=False),
+        g.nodes.to_df(explode=True, include_property_histories=False),
         pd.read_json(
             base_dir / "expected/dataframe_output/node_df_explode_no_hist.json"
         ),
     )
 
     # compare_df(
-    #     g.to_node_df(explode=True, convert_datetime=True),
+    #     g.nodes.to_df(explode=True, convert_datetime=True),
     #     pd.read_json(base_dir / "expected/dataframe_output/node_df_explode_datetime.json")
     # )
