+++ raphtory/src/python/graph/edges.rs
@@ -1,3 +1,4 @@
+use std::collections::HashMap;
 use crate::{
     core::ArcStr,
     db::{
@@ -21,10 +22,19 @@ use crate::{
             },
         },
         utils::PyTime,
+        utils::export::{extract_properties, get_column_names_from_props, create_row},
     },
 };
 use itertools::Itertools;
-use pyo3::{pyclass, pymethods, IntoPy, PyObject, Python};
+use pyo3::{prelude::*, pyclass, pymethods, IntoPy, PyObject, Python, PyResult, ToPyObject};
+use pyo3::prelude::PyModule;
+use pyo3::types::PyDict;
+use crate::core::Prop;
+use rayon::prelude::*;
+use rayon::iter::{IntoParallelIterator};
+use crate::db::api::view::internal::CoreGraphOps;
+use crate::prelude::NodeViewOps;
+
 
 /// A list of edges that can be iterated over.
 #[pyclass(name = "Edges")]
@@ -228,6 +238,98 @@ impl PyEdges {
         let edges = self.edges.clone();
         (move || edges.layer_names().map(|e| e.collect_vec())).into()
     }
+
+
+    /// Converts the graph's edges into a Pandas DataFrame.
+    ///
+    /// This method will create a DataFrame with the following columns:
+    /// - "src": The source node of the edge.
+    /// - "dst": The destination node of the edge.
+    /// - "layer": The layer of the edge.
+    /// - "properties": The properties of the edge. This column will be included if `include_edge_properties` is set to `true`.
+    /// - "property_histories": The history of the edge's properties. This column will be included if both `include_edge_properties` and `include_property_histories` are set to `true`.
+    /// - "update_history": The update history of the edge. This column will be included if `include_update_history` is set to `true`.
+    /// - "update_history_exploded": The exploded update history of the edge. This column will be included if `explode_edges` is set to `true`.
+    ///
+    /// Args:
+    ///     include_property_histories (bool): A boolean, if set to `true`, the history of each property is included, if `false`, only the latest value is shown. Ignored if exploded. Defaults to `false`.
+    ///     convert_datetime (bool): A boolean, if set to `true` will convert the timestamp to python datetimes, defaults to `false`
+    ///     explode (bool): A boolean, if set to `true`, will explode each edge update into its own row. Defaults to `false`
+    ///
+    /// Returns:
+    ///     If successful, this PyObject will be a Pandas DataFrame.
+    #[pyo3(signature = (include_property_histories=true, convert_datetime=false, explode=false))]
+    pub fn to_df(
+        &self,
+        include_property_histories: bool,
+        convert_datetime: bool,
+        explode: bool,
+    ) -> PyResult<PyObject> {
+        let mut column_names = vec![
+            String::from("src"),
+            String::from("dst"),
+            String::from("layer"),
+        ];
+        let edge_meta = self.edges.graph.edge_meta();
+        let is_prop_both_temp_and_const =
+            get_column_names_from_props(&mut column_names, edge_meta);
+
+        let mut edges = self.edges.explode_layers();
+        if explode == true {
+            edges = self.edges.explode_layers().explode();
+        }
+
+        let edge_tuples: Vec<_> = edges
+            .collect()
+            .into_par_iter()
+            .flat_map(|item| {
+                let mut properties_map: HashMap<String, Prop> = HashMap::new();
+                let mut prop_time_dict: HashMap<i64, HashMap<String, Prop>> = HashMap::new();
+
+                extract_properties(
+                    include_property_histories,
+                    convert_datetime,
+                    explode,
+                    &column_names,
+                    &is_prop_both_temp_and_const,
+                    &item.properties(),
+                    &mut properties_map,
+                    &mut prop_time_dict,
+                    item.start().unwrap_or(0),
+                );
+
+                let row_header: Vec<Prop> = vec![
+                    Prop::from(item.src().name()),
+                    Prop::from(item.dst().name()),
+                    Prop::from(item.layer_name().unwrap_or(ArcStr::from(""))),
+                ];
+
+                let start_point = 3;
+                let history = item.history();
+
+                create_row(
+                    convert_datetime,
+                    explode,
+                    &column_names,
+                    properties_map,
+                    prop_time_dict,
+                    row_header,
+                    start_point,
+                    history,
+                )
+            })
+            .collect();
+
+        Python::with_gil(|py| {
+            let pandas = PyModule::import(py, "pandas")?;
+            let kwargs = PyDict::new(py);
+            kwargs.set_item("columns", column_names)?;
+            let df = pandas.call_method("DataFrame", (edge_tuples,), Some(kwargs))?;
+            Ok(df.to_object(py))
+        })
+    }
+
+
 }
 
 impl Repr for PyEdges {
