+++ raphtory/src/python/utils/mod.rs
@@ -16,6 +16,7 @@ use pyo3::{exceptions::PyTypeError, prelude::*, types::PyDateTime};
 use std::{future::Future, thread};
 
 pub mod errors;
+pub(crate) mod export;
 
 /// Extract a `NodeRef` from a Python object.
 /// The object can be a `str`, `u64` or `PyNode`.
